refactor(clawdstrike-policy-event): extract finding_builders from edr/mod.rs
Move the detection finding builders (FindingRule, finding, ev/opt_ev,
identity-evidence stamping) and the supply-chain path/script/credential
predicates (is_install_phase, suspicious_script_reason, path_is_*,
path_looks_like_*, credential_kind_is_developer_secret,
credential_kind_from_secret, contains_any) into edr/finding_builders.rs.

Named finding_builders (not finding) to avoid colliding with the
existing detection/finding.rs, and NOT glob-re-exported. finding is used
by detection/{mod,supply_chain}, receipt/{mod,inputs/detection}; the path
predicates + ev/opt_ev by detection/supply_chain; credential_kind_from_
secret by the conversion cluster still in mod.rs. Those items are
re-exported pub(crate) at the edr root (private fns widened to pub(crate)
as needed). The finding_builders fns reach conversion-cluster extractors
(agent_id_field, tool_name_field, etc.) and util via super::. Pure
relocation; bodies unchanged.
